Remove validation on texts.
My reasons for this are somewhat abstract, but let's go through them.
Regex is the first to fall, because it would require a new type object
to be added to tasso to become self-describing. Other alternatives, like

    {
      type: 'regex',
      body: '[+-]?[1-9]\\d*(?:\\.\\d+)?(?:e[+-][1-9]\\d*)?',
      flags: 'i'
    }

are not just deeply clumsy but they mean that a model cannot be verified
by a metamodel because the language of regular expressions is famously
not regular. Compared to that clumsiness, just saying that user-space
logic is necessary for regex validation of passed strings just seems
cleaner.

This guides me towards the idea that there is no clear control structure
which corresponds to that validation anyway. Whether numbers should have
the integer flag becomes interesting, too, but I will perhaps leave that
in as a fringe case. But runtime tests on string length and regular
expression are so far from what TypeScript wants to do, that it does not
seem like a job for type objects.